Rename "caching" package to "cache"
- For consistency with "log" package, which also doesn't use the gerund form of the word
